+++ src/main/kotlin/io/github/e1turin/itmochan/entity/Picture.kt
@@ -3,12 +3,12 @@ package io.github.e1turin.itmochan.entity
 import java.io.File
 
 data class Picture(
-    val pictureId : Int,
+    val pictureId : Long,
     val name : String,
     val file : File,
 )
 
 data class PictureAttachments(
-    val commentId : Int,
-    val pictureId : Int,
+    val commentId : Long,
+    val pictureId : Long,
 )
