+++ src/main/kotlin/io/github/e1turin/itmochan/entity/Topic.kt
@@ -1,7 +1,7 @@
 package io.github.e1turin.itmochan.entity
 
 data class Topic(
-    val topicId : Int,
+    val topicId : Long,
     val name : String,
     val description : String,
 )
