+++ src/main/kotlin/io/github/e1turin/itmochan/entity/Trash.kt
@@ -3,8 +3,8 @@ package io.github.e1turin.itmochan.entity
 import java.time.LocalDateTime
 
 data class Trash(
-    val trashId : Int,
-    val commentId : Int,
+    val trashId : Long,
+    val commentId : Long,
     val reason : String?,
     val recycleDate : LocalDateTime,
 )
