+++ src/main/kotlin/io/github/e1turin/itmochan/entity/Video.kt
@@ -3,13 +3,13 @@ package io.github.e1turin.itmochan.entity
 import java.io.File
 
 data class Video(
-    val videoId : Int,
+    val videoId : Long,
     val name : String,
     val file : File,
 )
 
 data class VideoAttachments(
-    val commentId : Int,
-    val videoId : Int,
+    val commentId : Long,
+    val videoId : Long,
 )
